+++ R/tokenlist.R
@@ -180,6 +180,37 @@ tokenlist_filter <- function(x, dict, keep = FALSE) {
   new_tokenlist(out, lemma = lemma, pos = pos, unique_tokens = dict)
 }
 
+tokenlist_filter_function <- function(x, fn) {
+  if (!is_tokenlist(x)) {
+    rlang::abort("Input must be a tokenlist.")
+  }
+  
+  tokens <- get_tokens(x)
+  
+  keeps <- lapply(tokens, fn)
+  
+  out <- purrr::map2(tokens, keeps, ~.x[.y])
+  
+  lemma <- maybe_get_lemma(x)
+  if (!is.null(lemma)) {
+    lemma <- purrr::map2(lemma, keeps, ~.x[.y])
+    names(lemma) <- NULL
+  } else {
+    lemma <- NULL
+  }
+  
+  pos <- maybe_get_pos(x)
+  if (!is.null(pos)) {
+    pos <- purrr::map2(pos, keeps, ~.x[.y])
+    names(pos) <- NULL
+  } else {
+    pos <- NULL
+  }
+  
+  tokenlist(out, lemma = lemma, pos = pos)
+}
+
+
 tokenlist_apply <- function(x, fun, arguments = NULL) {
   if (!is_tokenlist(x)) {
     rlang::abort("Input must be a tokenlist.")